Doubly linked lists

+++ 0x17-doubly_linked_lists/5-get_dnodeint.c
@@ -3,6 +3,7 @@
 /**
  * get_dnodeint_at_index - Locates a given node of
  *                        a dlistint_t linked list.
+ *
  * @head: A pointer to the head of the listint_t list.
  * @index: The index of the node to locate - indices start at 0.
  *
@@ -12,7 +13,7 @@
 
 dlistint_t *get_dnodeint_at_index(dlistint_t *head, unsigned int index)
 {
-	unsigned int count;
+	long int count;
 
 	if (head == NULL)
 	{
